Remove stalkers from spines.
It is weird for a spine to be able to communicate with a device
directly. It means you can't migrate a bridge to a different spine.
Instead, it makes more sense for there to be a mode where you aren't in
a leaf spine architecture, but your'e just a singular node.